Refactored server.js functionality into Router files and updated gitignore to ignore IDE preferences
