+++ src/HelloWorldRouter.js
@@ -1,0 +1,16 @@
+// @flow
+import AbstractRouter from './AbstractRouter';
+
+class HelloWorldRouter extends AbstractRouter {
+
+    constructor() {
+        super('GET', '/', false);
+    }
+
+    async content(req: Request): Promise<any> {
+        return 'hello, world\n';
+    }
+
+}
+
+export default new HelloWorldRouter().router;
